working task list by user on modal
