+++ src/Components/MDropDown.js
@@ -1,56 +1,61 @@
-import React, { useState } from "react";
-import {Dropdown, Button} from "react-bootstrap";
+// import React, { useState, useEffect } from "react";
+// import {Dropdown, Button} from "react-bootstrap";
 
-export default function MDropDown() {
-    const [tasks, setTasks] = useState([]); 
-    const [assignedContractor,setAssignedContractor] = useState("")
-//fetch to get task by user
-const usertask =() => {
-  fetch(`https://ezcontractz-backend.herokuapp.com/tasks/${localStorage.getItem("UserName")}`)
-    .then((res) => res.json())
-    .then((data) => {
-        if (data.error) {
-          alert(data.error);
-        } else {
-          console.log(data);
-          setTasks(data.task);
-        }        
-    })
-// //put to update the assigned contractor to the task   
-// const assignContractor = () => {
-// fetch(`https://ezcontractz-backend.herokuapp.com/tasks/${localStorage.getItem("UserName")}`,{
-//           method: "PUT",
-//           credentials: "include",
-//           headers: {
-//             "Content-Type": "application/json",
-//           },
-//         })
-//           .then((res) => res.json())
-//           .then((data) => {
-//             if (data.error) {
-//               alert(data.error);
-//             } else {
-//               console.log(data);
-//               setAssignedContractor(assignedContractor);
-//             }
-//           });
-//         }
-return (
+// export default function MDropDown() {
+//     const [tasks, setTasks] = useState([]); 
+//     const [assignedContractor,setAssignedContractor] = useState("")
+//     useEffect(() => {
+//         //fetch to get task by user
+//     const usertask =() => {
+//     fetch(`https://ezcontractz-backend.herokuapp.com/tasks/${localStorage.getItem("UserName")}`)
+//       .then((res) => res.json())
+//       .then((data) => {
+//           if (data.error) {
+//             alert(data.error);
+//           } else {
+//             console.log(data);
+//             setTasks(data.task);
+//           }        
+//       })
+//     }
+//     usertask()
+//     }, [])
+// //fetch to get task by user
+// const usertask =() => {
+//   fetch(`https://ezcontractz-backend.herokuapp.com/tasks/${localStorage.getItem("UserName")}`)
+//     .then((res) => res.json())
+//     .then((data) => {
+//         if (data.error) {
+//           alert(data.error);
+//         } else {
+//           console.log(data);
+//           setTasks(data.task);
+//         }        
+//     })
+// // //put to update the assigned contractor to the task   
+// // const assignContractor = () => {
+// // fetch(`https://ezcontractz-backend.herokuapp.com/tasks/${localStorage.getItem("UserName")}`,{
+// //           method: "PUT",
+// //           credentials: "include",
+// //           headers: {
+// //             "Content-Type": "application/json",
+// //           },
+// //         })
+// //           .then((res) => res.json())
+// //           .then((data) => {
+// //             if (data.error) {
+// //               alert(data.error);
+// //             } else {
+// //               console.log(data);
+// //               setAssignedContractor(assignedContractor);
+// //             }
+// //           });
+// //         }
+// return (
     
-    <div className = "cg">
-    <Button variant="primary" classname="assignCButton" onClick={() => usertask()}>
-                    Assign Contractor
-                  </Button>
-    <Dropdown.Menu show> {
-        tasks.map((task, index) => { 
-         return(   <div key={index}>           
-            <Dropdown.Item eventKey="1">{task.title}</Dropdown.Item>
-            <Dropdown.Divider /> 
-            </div>    
-         )          
-        }         
-        )
-    }</Dropdown.Menu>
-    </div>
-)
-}}
+//     // <div className = "cg">
+    
+    
+//     // </div>
+// )
+// }}
